Load all records into memory as they arrive
