+++ app/adapters/application.js
@@ -6,7 +6,9 @@ function setupDb() {
   let localDb = new PouchDB(config.emberPouch.localDb);
 
   if (config.emberPouch.remoteDb) {
-    let remoteDb = new PouchDB(config.emberPouch.remoteDb);
+    let remoteDb = new PouchDB(config.emberPouch.remoteDb, {
+      ajax: { timeout: 20000 }
+    });
 
     localDb.sync(remoteDb, {
       live: true,
@@ -21,5 +23,15 @@ export default Adapter.extend({
   init() {
     this._super(...arguments);
     this.set('db', setupDb());
+  },
+
+  // Load all records into memory as they arrive.
+  unloadedDocumentChanged(obj) {
+    let store = this.get('store');
+    let recordTypeName = this.getRecordTypeName(store.modelFor(obj.type));
+
+    this.get('db').rel.find(recordTypeName, obj.id).then(function(doc) {
+      store.pushPayload(recordTypeName, doc);
+    });
   }
 });
