--- app/adapters/tag.js
@@ -1,3 +1,0 @@
-import AutoLoadAdapter from './auto-load';
-
-export default AutoLoadAdapter.extend();
